Move Kithe::Indexable.settings to Kithe.indexable_settings
Necessary to play well with Rails dev-mode class reloading. Since the Kithe::Indexable class gets reloaded by Rails sometimes, when it does settings set on Kithe::Indexable.settings in an initializer were lost.

We should keep all settings in the 'Kithe' module, which does not get reloaded. This is actually what Rails guide recommends, although it doesn't explain why.
https://guides.rubyonrails.org/engines.html#configuring-an-engine
