+++ app/indexing/kithe/indexable.rb
@@ -9,7 +9,7 @@ module Kithe
   # For a complete overview, see the [Guide Documentation](../../../guides/solr_indexing.md)
   #
   # The Solr instance to send updates to is global configuration:
-  #     Kithe::Indexable.settings.solr_url = "http://localhost:8983/solr/collection_name"
+  #     Kithe.indexable_settings.solr_url = "http://localhost:8983/solr/collection_name"
   #
   # To configure how a model is mapped to a Solr document, you create a `Kithe::Indexer` sub-class, which
   # can use our obj_extract method, as well as any other traject indexer code.
@@ -52,98 +52,6 @@ module Kithe
   module Indexable
     extend ActiveSupport::Concern
 
-    class IndexableSettings
-      attr_accessor :solr_url, :writer_class_name, :writer_settings,
-                    :model_name_solr_field, :solr_id_value_attribute, :disable_callbacks
-      def initialize(solr_url:, writer_class_name:, writer_settings:,
-                     model_name_solr_field:, solr_id_value_attribute:, disable_callbacks: false)
-        @solr_url = solr_url
-        @writer_class_name = writer_class_name
-        @writer_settings = writer_settings
-        @model_name_solr_field = model_name_solr_field
-        @solr_id_value_attribute = solr_id_value_attribute
-      end
-
-      # Use configured solr_url, and merge together with configured
-      # writer_settings
-      def writer_settings
-        if solr_url
-          { "solr.url" => solr_url }.merge(@writer_settings)
-        else
-          @writer_settings
-        end
-      end
-
-      # Turn writer_class_name into an actual Class object.
-      def writer_class
-        writer_class_name.constantize
-      end
-
-      # Instantiate a new writer based on `writer_class_name` and `writer_settings`
-      def writer_instance!(additional_settings = {})
-        writer_class.new(writer_settings.merge(additional_settings))
-      end
-    end
-
-    # Global Kithe::Indexable settings, actually a Kithe::Indexable::Settings
-    # object, but you will generally use it as a simple value object with getters
-    # and setters.
-    #
-    # * solr_url: Where to send to Solr when indexing, the base url
-    #
-    #     Kithe::Indexable.settings.solr_url = "http://localhost:8983/solr/collection_name"
-    #
-    # * model_name_solr_field: If you'd like a custom solr field to store model class name in.
-    #
-    #     Kithe::Indexable.settings.model_name_solr_field = "my_model_name_field"
-    #
-    # * solr_id_value_attribute: What attribute from your AR models to send to Solr
-    #   `id` uniqueKey field, default the AR `id` pk, you may wish to set to `friendlier_id`.
-    #
-    # * writer_settings: Settings to be passed to the Traject writer, by default a
-    #   Traject::SolrJsonWriter. To maintain the default settings, best to merge
-    #   your new ones into defaults.
-    #
-    #       Kithe::Indexable.settings.writer_settings.merge!(
-    #         # by default we send a softCommit on every update, maybe you
-    #         # want not to?
-    #         "solr_writer.solr_update_args" => {}
-    #         # extra long timeout?
-    #         "solr_writer.http_timeout" => 100
-    #       )
-    #
-    # * writer_class_name: By default Traject::SolrJsonWriter, but maybe
-    #   you want to set to some other Traject::Writer. The writer Kithe::Indexable
-    #   will send index add/remove requests to.
-    #
-    #       Kithe::Indexable.settings.writer_class_name = "Traject::SomeOtherWriter"
-    #
-    # * disable_callbacks: set to true to globally disable automatic after_commit
-    mattr_accessor :settings do
-      # set up default settings
-      IndexableSettings.new(
-        solr_url: "http://localhost:8983/solr/default",
-        model_name_solr_field: "model_name_ssi",
-        solr_id_value_attribute: "id",
-        writer_class_name: "Traject::SolrJsonWriter",
-        writer_settings: {
-          # as default we tell the solrjsonwriter to use no threads,
-          # no batching. softCommit on every update. Least surprising
-          # default configuration.
-          "solr_writer.thread_pool" => 0,
-          "solr_writer.batch_size" => 1,
-          "solr_writer.solr_update_args" => { softCommit: true },
-          "solr_writer.http_timeout" => 3,
-          "logger" => Rails.logger,
-
-          # MAYBE? no skippable exceptions please
-          # "solr_writer.skippable_exceptions" => []
-        },
-        disable_callbacks: false
-      )
-    end
-
-
     # Set some indexing parameters for the block yielded. For instance, to batch updates:
     #
     #     Kithe::Indexable.index_with(batching: true)
@@ -170,12 +78,12 @@ module Kithe
 
     # Are automatic after_commit callbacks currently enabled? Will check a number
     # of things to see, as we have a number of places these can be turned on/off.
-    # * Globally in `Kithe::Indexable.settings.disable_callback`
+    # * Globally in `Kithe.indexable_settings.disable_callback`
     # * On class or instance using class_attribute `kithe_indexable_auto_callbacks`
     # * If no kithe_indexable_mapper is configured on record, then no callbacks.
     # * Using thread-current settings usually set by .index_with
     def self.auto_callbacks?(model)
-      !Kithe::Indexable.settings.disable_callbacks &&
+      !Kithe.indexable_settings.disable_callbacks &&
         model.kithe_indexable_auto_callbacks &&
         model.kithe_indexable_mapper &&
         !ThreadSettings.current.disabled_callbacks?
