+++ app/indexing/kithe/indexable/record_index_updater.rb
@@ -68,7 +68,7 @@ module Kithe
       # Could be an explicit writer passed into #initialize, or a current thread-settings
       # writer, or a new writer created from global settings.
       def writer
-        @writer ||= ThreadSettings.current.writer  || Kithe::Indexable.settings.writer_instance!
+        @writer ||= ThreadSettings.current.writer  || Kithe.indexable_settings.writer_instance!
       end
 
       # Is this record supposed to be represented in the solr index?
